+++ context/app/routes_main.py
@@ -1,5 +1,6 @@
 from flask import (Blueprint, render_template, abort, current_app,
-                   session, flash, request, get_flashed_messages)
+                   session, flash, request, get_flashed_messages,
+                   redirect, url_for)
 
 from yaml import safe_load as load_yaml
 
@@ -39,6 +40,10 @@ def details(type, uuid):
     client = _get_client()
 
     entity = client.get_entity(uuid)
+    actual_type = entity['entity_type'].lower()
+    if type != actual_type:
+        return redirect(url_for('routes.details', type=actual_type, uuid=uuid))
+
     # TODO: These schemas don't need to be reloaded per request.
     with open(current_app.root_path + f'/schemas/{type}.schema.yaml') as type_schema_file:
         type_schema = load_yaml(type_schema_file)