+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## v0.0.14 - In progress
+### Added
+- Redirect to the correct entity type, if we've landed on the wrong one.
 
 ## [v0.0.13](https://github.com/hubmapconsortium/portal-ui/tree/v0.0.13) - 2020/05/03
 ### Added